# 🦀 Rust implementation