--- a/.gitignore
+++ b/.gitignore
@@ -1,0 +1,4 @@
+
+rust-enjoyer/target/
+
+rust-enjoyer/Cargo.lock