@Vectorized 2023-11-25
Ok, letting it rest for a while then merging in a day. This is beautiful.