@Vectorized 2023-11-26
I reran the latest script and it seems to not be finding the selector.

```
cargo run --release "someContext" "(uint256)" "0f2c1eff"

AVX2 enabled
Function name: someContext
Function params: (uint256)
Target selector:  0xf2c1eff
Starting mining with 35 threads.
0 hashes done.
10276044800 hashes done.
20552089600 hashes done.

```