@Vectorized 2023-11-26
The commit "switch to array iters" at aff539b resulted in 2x slower code on my machine. O.O

```
time cargo run --release "someContext" "(uint256)" "0f2c1eff";
```

went from 9.6s to 21s.