@nascent-chris 2023-11-26
> The commit "switch to array iters" at [aff539b](https://github.com/Vectorized/function-selector-miner/commit/aff539b220982760cbc5ed8ef6ff0abfcb26048a) resulted in 2x slower code on my machine. O.O
> 
> ```
> time cargo run --release "someContext" "(uint256)" "0f2c1eff";
> ```
> 
> went from 9.6s to 21s.

did it find the same result?

I ran `cargo bench` between changes to be confident it didn't have an impact.  I'm wondering if it's related to the above issue of different values for `num_threads` causing misses for some reason

even stranger, the most recent commit changed the behavior :thinking: