@nascent-chris 2023-11-26
> The commit "switch to array iters" at [aff539b](https://github.com/Vectorized/function-selector-miner/commit/aff539b220982760cbc5ed8ef6ff0abfcb26048a) resulted in 2x slower code on my machine. O.O
> 
> ```
> time cargo run --release "someContext" "(uint256)" "0f2c1eff";
> ```
> 
> went from 9.6s to 21s.

on my machine currently:

`"someContext" "(uint256)" "0f2c1eff" 16`
5.495s

`"someContext" "(uint256)" "0f2c1eff" 32`
4.945s

`"someContext" "(uint256)" "0f2c1eff" 64`
2.469s