@nascent-chris 2023-11-26
> The commit "switch to array iters" at [aff539b](https://github.com/Vectorized/function-selector-miner/commit/aff539b220982760cbc5ed8ef6ff0abfcb26048a) resulted in 2x slower code on my machine. O.O
> 
> ```
> time cargo run --release "someContext" "(uint256)" "0f2c1eff";
> ```
> 
> went from 9.6s to 21s.

Here's the result of `cargo bench compute` across the two branches for me

With iterators, most recent:
```
compute                 time:   [412.06 ns 412.27 ns 412.50 ns]
```

Before iterators, commit e5630a5c48de80444322546fd2ad200425ec22a0
compute                 time:   [427.81 ns 430.60 ns 433.09 ns]
```
It varies a good amount between runs (sometimes very high sometimes lower), but this is typical results on my machine