@nascent-chris 2023-11-26
I've reverted from the iterators -- let me know if you see a difference between the two implementations

I have a feeling it gets optimized to be very similar code, based on on my experienced performance impact and messing around a bit on goldbolt https://godbolt.org/z/3brM57c8r