@Vectorized 2023-11-26
The latest commit at cbc33ef fixes it for me

```
git checkout -- .;
git checkout e2ca8c8;
echo "[build]
rustflags = [\"-C\", \"target-cpu=native\"]
" > .cargo/config.toml;

cargo build --release;
time cargo run --release "someContext" "(uint256)" "0f2c1eff" 36;

AVX2 enabled
Function name: someContext
Function params: (uint256)
Target selector: 0xff1e2c0f
Starting mining with 36 threads.
0 hashes done.
301989888 hashes done.
603979776 hashes done.
Function found: someContext665728898(uint256)

real	0m9.208s
user	5m9.176s
sys	0m0.032s
```

With 72+ threads, it sometimes get faster, but sometimes misses the hit at `665728898`.