@nascent-chris 2023-11-26
> e2ca8c8

e2ca8c82b517b4db859600088c69cf20d5346041 was the commit before I reverted the iterators, in case this wasn't a type-o