@Vectorized 2023-11-26
Sorry, i've made a typo.

With iterators e2ca8c8:
```
root@C.7523033:~/function-selector-miner/rust-enjoyer$ git checkout -- .;
root@C.7523033:~/function-selector-miner/rust-enjoyer$ git checkout e2ca8c8;
Previous HEAD position was cbc33ef revert iterator refactor
HEAD is now at e2ca8c8 PR comments
root@C.7523033:~/function-selector-miner/rust-enjoyer$ echo "[build]
> rustflags = [\"-C\", \"target-cpu=native\"]
> " > .cargo/config.toml;
root@C.7523033:~/function-selector-miner/rust-enjoyer$  
root@C.7523033:~/function-selector-miner/rust-enjoyer$ cargo build --release;
   Compiling rust-enjoyer v0.1.0 (/root/function-selector-miner/rust-enjoyer)
    Building [========================>  ] 23/24: rust-enjoyer(bin)                           
    Finished release [optimized] target(s) in 6.43s
root@C.7523033:~/function-selector-miner/rust-enjoyer$ time cargo run --release "someContext" "(uint256)" "0f2c1eff" 36;
    Finished release [optimized] target(s) in 0.04s
     Running `target/release/rust-enjoyer someContext '(uint256)' 0f2c1eff 36`
AVX2 enabled
Function name: someContext
Function params: (uint256)
Target selector: 0xff1e2c0f
Starting mining with 36 threads.
0 hashes done.
301989888 hashes done.
603979776 hashes done.
Function found: someContext665728898(uint256)

real	0m21.486s
user	12m6.342s
sys	0m0.068s
```

Without iterators cbc33ef:
```
root@C.7523033:~/function-selector-miner/rust-enjoyer$ git checkout -- .;
root@C.7523033:~/function-selector-miner/rust-enjoyer$ git checkout cbc33ef;
Previous HEAD position was e2ca8c8 PR comments
HEAD is now at cbc33ef revert iterator refactor
root@C.7523033:~/function-selector-miner/rust-enjoyer$ echo "[build]
> rustflags = [\"-C\", \"target-cpu=native\"]
> " > .cargo/config.toml;
root@C.7523033:~/function-selector-miner/rust-enjoyer$  
root@C.7523033:~/function-selector-miner/rust-enjoyer$ cargo build --release;
   Compiling rust-enjoyer v0.1.0 (/root/function-selector-miner/rust-enjoyer)
    Finished release [optimized] target(s) in 6.00s
root@C.7523033:~/function-selector-miner/rust-enjoyer$ time cargo run --release "someContext" "(uint256)" "0f2c1eff" 36;
    Finished release [optimized] target(s) in 0.04s
     Running `target/release/rust-enjoyer someContext '(uint256)' 0f2c1eff 36`
AVX2 enabled
Function name: someContext
Function params: (uint256)
Target selector: 0xff1e2c0f
Starting mining with 36 threads.
0 hashes done.
301989888 hashes done.
603979776 hashes done.
Function found: someContext665728898(uint256)

real	0m9.178s
user	5m7.638s
sys	0m0.028s
```