@nascent-chris 2023-11-26
very interesting!

roger that, we'll leave it without iterators

I wonder why there's such a big difference :thinking: